requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
googleapis-common-protos==1.67.0
greenlet==3.1.1
groq==0.18.0
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-google-community==2.0.4
langchain-groq==0.2.4
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langchainhub==0.1.21
langgraph==0.2.73
langgraph-checkpoint==2.0.15
langgraph-sdk==0.1.51
langsmith==0.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.27.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.63.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==24.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.7.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
setuptools==65.5.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
streamlit==1.42.2
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
yarl==1.18.3
zstandard==0.23.0
